plot: add typing rules to sherpa.plot/sherpa.astro.plot
These changes mainly point out that our class structure doesn't
follow "classical" C++-style rules (e.g. our method over-rides
often drastically change calling conventions, and I've finally
managed to get mypy to mention the Liskov Substitution Principle in
an error message) and we have a few cases where we have to help
mypy follow the logic. This is mainly because we require that the
prepare method is called before the plot method, but do not actually
check that it has been. These are all technically issues with our
code that mypy is telling us about: we just need to decide whether
we want to do anything about it.

For the plot/prepare calls I add `# type: ignore[override]` to
hide the mypy warnings and to point out to the reader where we
have issues. Unfortunately this seems to reduce the effectiveness
of type checking on these routines.

There are some follow-on changes to typing rules - e.g. to the
Data classes - but these are not "complete" as that would be a lot
of extra work for little gain here.