plot: sort out xerr for histogram-style plots
So, avoid creating an xerr field for some histogram plots (those
that have a yerr field), but add it as a property - defined to
be (xhi - xlo) / 2, so the half-width of a bin - so that users
switching between "plot" and "histogram" plots do not have to
change their code (although they probably should). This follows
the x property - defined as (xlo + xhi) / 2 - but it's interesting
to see that it isn't added to the same class, but a child of
BaseHistogram/HistogramPlot. This actually has down-stream
consequences since the residual classes for Data1D data end up
inheriting from the Model rather than Data plot class, which has
no "x error" concept, and makes future plans more interesting.

It also turns out the backend code for histogram-style plots
ignores the xerr setting it is sent, and instead calculate it
directly, so we really should just remove it!