Tests: check out xerr handling of Data1D/1DInt/PHA
The xerr field of the plot objects is poorly documented, leading
to issues like #1817.

Add some basic checks to ensure we capture the current behaviour,
as I plan to change how xerr is handled for the Data1DInt/PHA
classes.

These tests show that there is something screwy going on with
xerr handling for DataPHA data when analysis=wavelength.