plot: error out if prepare has not been called
Error out if the prepare method has not been called for the plot
objects. We could just allow the "empty" graph to be created, but
it's not obvious that makes sense. The way it's done is a bit
noisy - e.g. the need to call check_not_none to strip out the
optional nature of a value - but is not hugely invasive.

This does change the behaviour of the system slightly - e.g.

- the ui layer with replot=True is likely to cause confusion,
  but to be honest this setting is confusing anyway

As an example, in 4.16

  >>> from sherpa.ui.utils import *
  >>> clean()
  >>> load_arrays(1, [1, 10], [50, 200])
  >>> plot_data(replot=True)

will create an "empty" plot (the exact behaviour depends on the plot
backend).

We now get

  >>> plot_data(replot=True)
  ...
  PlotErr: prepare has not been called

There's also some extra validation done in some of the prepare
methods to ensure the arguments are "correct" (e.g. DataPHA or
an instrument response). This is hoped to help out once we improve
the typing statements.

So, with 4.16 we have:

  >>> from sherpa import ui
  >>> ui.clean()
  >>> ui.plot_data(replot=True)
  ... will create a plot with matplotlib
  >>> ui.plot_data(replot=False)
  ...
  IdentifierErr: data set 1 has not been set
  >>> ui.get_data_plot(recalc=True)
  ...
  IdentifierErr: data set 1 has not been set
  >>> ui.get_data_plot(recalc=False)
  <sherpa.plot.DataPlot object at 0x...>

With this PR we get similar behaviour except that the first call
(plot_data with replot=True) errors out with

  PlotErr: prepare has not been called

but the other calls remain the same.