plot: make more use of super
Change from hard-coding the class name to use super **when**
this is easy to do. We have some classes that end up importing
both the ModelPlot and HistogramPlot classes, and it's not
obvious what we really want these classes to do in a
cooperative setting, so these are left alone.