data: xerr is now half the bin width for Data1DInt/PHA
The xerr Data field is not well defined for Sherpa, since we
do not support uncertainties on the independent axis. It appears
to have been thrown in as a way to indicate the width of a bin
for histogram data (so Data1DInt and DataPHA). However, it was
returning the full-width of each bin, which is confusing and
leads to problems like issue #1817.

There is also no recognition of setting the units to "wavelength"
for the DataPHA case (issue #1985).

This commit ensures that Data1DInt and DataPHA objects return the
half-width of each bin, and handles the units setting for DataPHA.
This therefore fixes #1985. I do not want to claim that #1817 is
addressed since although this change technically addresses the
issue, we really should handle the residual-style plots for
integrated data "better" (i.e. do not treat as unbinned).

There is a choice of how to handle this at the plotting backend
stage - we can

  - keep the old interface and just only ever send in half the old
    "full width" value (i.e. keep the backend code as is with it's
    "we mean xerr to actually be twice the yerr setting")

  - change the backend so that xerr means the same as yerr

I have elected to go with the latter.

The more I look at the xerr handlnig the more I think we should just
drop it, but we would need to "properly" solve #1817 first. This
is in a follow-up PR (where we extend "histogram" support to the
residual-style plot classes), but I am still not sure it is worth
the effort.

There is also a small fix to docstrings where we had an erroneous
" character. As part of this a few other docstrings were tweaked.