plot: fix an error check for an uncommon ARFPlot issue
The sherpa.astro.plot.ARFPlot would raise a PlotErr with a 'notpha'
label, but this is actually an IOErr case, so switch it.

This was found when adding tests just to check the error handling
of the PHA-related routines - that is, what happens when sent a
non-PHA object? These tests are treated as regression tests as the
plan is to standardize the behaviour.

The sherpa.astro.plot module has also seen some small cleanups, mainly
not importing symbols from numpy directly so you have to use np.<name>
to access them (and so know where the symbol comes from).