Tests: add suggestions from review
Ensure several conditions suggested by @hamogu are included.

Of note are the behaviour of the string output for

- parameters
- using "binary" functions which are not the standard operators
  we have been checking for

In this case some of the logic for deciding whether to add brackets
or not are off, since with

  numpy.funcname(a, b)

we don't need to worry about the precedences for a and b when deciding
whether to add () around them. However, the current code does not
recognize this case (as it was written before #1653 was merged). For
now we leave this behaviour in, but at least have checks to note
if we ever address this.