models: move precedence code into sherpa.models.op
These routines are not generic enough to go into sherpa.utils,
at least not at the moment, so move them into sherpa.models.op (it
is helpful to have them in their own module).

Following review comments, the lprec and rprec dictonaries are
no module-level, and so can be inspected and changed by the user
if necessary.