parameter: pylint cleanup
Change from "  % ..." and .format statements to r-strings.
Note that there is an assert that has not been changed: this
assert should really be removed but I am wary about doing that at
this time (in case some other code relies on this behaviour,
which I think is unlikely).

Use lazy % formatting for warning statements.

I have also taken the time to switch to using np rather than numpy.