parameter: improve string output for numpy binary functions
The bracket-cleaning was developed a while ago, before support
for "generic" NumPy binary functions was added for parameters in
PR #1653. This lead to adding extra brackets when not needed,
because the new code was for prefix operators - where we do not
need to add brackets around the lhs and rhs terms based on
parameter precedences - whereas the existing code only supported
or cared about infix operators.

The change in #1653 could have been done in terms of prefix vs
infix, rather than sending around the formatting we need to use
(which is more expressive, but may be too expressive). We leverage
this code to determine whether we have a prefix case - that is,
strformat begins with '{opstr}(' - in which case we do not need
to worry about the preferences.

This *could* be applied to the models, but we currently do not
suppport "generic" NumPy binary fnuctions for them (although we
should).