Tests: improve coverage of parameter module
When working on some recent changes to the model and parameter code,
such as #1972 and #1974, I noticed some code paths that were not
being checked. So this attempts to fix these (although it does not
guarantee to do so since there's a cost-benefit trade off here).

Writing these tests lead to noting issues

- #1992
- #1993

but these are not addressed here (but any changes to these issues
is likely to be caught by these tests, which was the point in
writing them!).