plot: RMFPlot can now be controlled with the energies field
When energies is None (the default) we automatically select the
energies to use in the plot (as before), and the number of lines
is controlled by the n_lines field.

When it is set then it is used to determine what lines to draw
(after filtering by the energy grid of the response).

The API is not ideal, particularly for the UI layer, but we have
issues with setting this sort of data for plots as they are not
realy preferences (used by the plot call) but settings for the
prepare call (see my attempts to control background plotting for
PHA data; issue #889 is somewhat related but not a perfect match
to this issue).