XSPEC: bump switch parameter for supported models for 12.14.0
There are a large number of changes in XSPEC 12.14.0 related to
improving thermal-plasma models (APEC and Mekal). One of the
changes is supporting a "use SPEX for interpolation" setting,
which is the ability to set switch to 3, and to change the
default setting from 1 to 2 [*].

There is no support for the new models yet.

[*] I have a question to the XSPEC help desk to question whether
    the cemekl switch parameter should be set to a default of 2
    as HEASOFT 6.33.1 / XSPEC 12.14.0b still leaves it at 1 even
    though cempow (which should be the same) has it set to 2