XSPEC: improve helper code for binding XSPEC models
Relatively recently XSPEC has started to provide models using
the double-precision FORTRAN interface, ni particular

  ismabs       31   0.01    1.e6      F_ismabs     mul  0
  ismdust       3   0.01    1.e6      F_ismdust     mul  0
  olivineabs       2   0.01    1.e6      F_olivineabs     mul  0

The helper code that can be used to help generate either the
XSPEC code for a new release, or for binding user models, had
not been updated to suppport this case. This commit fixes the
issue. Unfortunately this is not as generic as we'd like as the
names and functions don't quite match (on both the XSPEC and
Sherpa sides) between C, C++, and FORTRAN interfaces for the
single/double precision cases.