XSPEC: add scripts/update_xspec_functions
Add a script to make updating the XSPECMODELFCTxxx lines in
sherpa/astro/src/_xspec.cc a bit easier. Unfortunately we can
not just automatically include the output of this script into
the file because the script can not deal with version constraints
- that is, this model was changed in 12.13.0. It is mainly
useful if you have a lot of symbols to add (for example, as is
the case in XSPEC 12.14.0).