XSPEC: add 50 new models from XSPEC 12.14.0
This adds quite the number of models. Unfortunately it's not
obvious how to link between the models, but that's a common
issue with trying to document XSPEC models.

The new models are:

  XSbcempow, XSbcheb6, XSbcie, XSbcoolflow, XSbcph,
  XSbequil, XSbexpcheb6, XSbgadem, XSbgnei, XSbnei,
  XSbsnapec, XSbnpshock, XSbpshock, XSbsedov, XSbvcempow,
  XSbvcheb6, XSbvcie, XSbvcoolflow, XSbvcph, XSbvequil,
  XSbvexpcheb6, XSbvgadem, XSbvgnei, XSbvnei, XSbvnpshock,
  XSbvpshock, XSbvsedov, XSbvvcie, XSbvvgadem, XSbvvgnei,
  XSbvvnei, XSbvvnpshock, XSbvvpshock, XSbvvsedov, XSbvvwdem,
  XSbvwdem, XSbwdem, XScempow, XScheb6, XScie, XScoolflow,
  XSeebremss, XSexpcheb6, XSvcempow, XSvcheb6, XSvcie,
  XSvcoolflow, XSvexpcheb6, XSvvcie, XSvvgadem

Note that some of these are replacements for existing models,
such as XScempow and XSvcempow being replacements for
XScemelk and XScevmkl, but we have not indicated this in
any way (other than perhaps linking the different classes
in the 'See Also' section) since the models are listed with
different function names (even if internally they end up
being an alias, we can not tell that from the model.dat file).

The documentation has been updated to point out XSPEC 12.14.0
is supported, as well as adding in a few model classes that
had been missed in previous updates.