XSPEC: rework the layout of the _xspec.cc module
This uses the new scripts/update_xspec_functions.py script to
set up the model definitions. This has the advantage that the
output now matches the ordering of the model.dat file (here
taken from HEASOFT 6.32) rather than being something that has
just developed over time.

The intent is that

- there is no functional change
- it will be easier to see when routines get added and removed,
  as we can more-easily compare against the model.dat file
  (e.g. when adding support for XSPEC 12.14.0)

Note that sherpa/astro/xspec/__init__.py already matches
the order of model.dat (that is, the classes are in the same
order as they appear in model.dat).