XSPEC: fix up documentation
Update the documentation to match the output from #1972 (the
removal of excess brackets from the name of a model).

Unfortunately the XSPEC model is hard to include in testing,
with the doc tests, as there is too much implicit state
required (e.g. files, setting of environment variables, ...)
so it's easy to get out of date.