SCRIPTS: improve XSPEC parameter checking
I had not been checking the frozen status of parameter values.
This should fix that.

It also tweaks how XSPEC model definitions are displayed by
add_xspec_model.py, but this is just a style change (adding a
space after each element in a tuple) and not a functional change.