XSPEC: normalize the norm parameter
The normalization parameter for XSPEC additive models is added
by Sherpa since XSPEC doesn't have this (it is an "implicit"
parameter handled by XSPEC). This has lead to small deviations
in the parameters - mainly in what the hard_max is (is it
10^24 or hugeval?) - over time.

So, we now automatically add the norm parameter **only if**
the pars list sent to the constructor does not end with a
norm parameter. This lets us avoid defining the parameter
explicitly for virtually all additive models, while letting
us over-ride it when necessary. These over-ride cases are
ones where the XSPEC documentation makes it clear that the
normalization should not be thawed, but undortunately this
knowledge is not encoded in the model.dat file soi it has to
be handled manually.

There's also some pylint clean up

- f-string/warning-message changes
- raise .... from None to simplify backtraces
- no 'if .. return else .. return'
- fixed a logical error in a fold method (it didn't take self
  as the first argument, but all it did was call pass anyway)

I have also reworked most of the additive model initalizations
to set up a pars tuple which is then sent to __init__, to avoid
long lines. This is unfortunately "noise" when coming to review
this commit, since it should not change behaviour.