XSPEC: use correct capitalization of Redshift for parameters
Change the case of the redshift parameter to match the XSPEC
model.dat file: sometimes it's redshift and sometimes it's
Redshift. This has no impact on users (other than visually when
looking at the string output of a model) since the parameter
access is case insensitive. The main reason I am doing this is
to reduce the number of "false positives" that
./scripts/check_xspec_update.py reports.