plot: rationalize the residual plot classes
The residual-style plots (resid, ratio, delchi but *not* chisqr)
have the same structure, and really only differ on how y and dy
are calculated. So intoduce a class that encapsulates this logic
in a more-structured way than the old approach - which had moved
the "calculate y" into a separate routine but each class had a
different name.

The chisqr plot doesn't quite match, so this is left as is. There
is a note about whether we want to draw y or x errors (I think
dy doesn't make sense).