CI: use macOS ARM build rather than Intel
The macOS ARM installations (-13 and -14) do not provide
miniconda, so we install it ourselves.

Note that we do not provide a macOS arm conda package for
xspec-modelsonly at this time, so we take the package from the
CIAO build in this situation, and hence we use an older XSPEC
version (12.13.1 rather than 12.14.0). This is hopefully a
temporary solution.