CI: use macOS ARM and Intel builds
Run the macOS Intel on macos-13, which is the last version
to support Intel without switching to the "large" version,
and macOS ARM on macos-14.

At present, since we have to rely in the CXC "ciao" channel
for xspec-modelsonly on ARM we are forced to use an older
XSPEC version for the ARM test.