plot: contour-style residual plots
Update the contour (Data2D) code to match the recent changes for
residual plots for plot (Data1D) and histogram (Data1DInt) cases.

There is less use in this BaseResidualContour class since we only
have the RatioContour and ResidContour classes, but it seems
sensible to apply the same refactoring as we did for the 1D cases.