IO: break up PHA reading
There should be no functional change [*], it just makes the code
easier to read and easier to type.

[*] if you are using some of the internal routines from the
    sherpa.astro.io.pyfits_backend module then there are
    user visible changes, but these routines are marked internal
    (i.e. start with the underscore character).

The idea is (for some of these routines) is to make it more
obvious when we haven't read in the key or column by returning
None rather than some sort of a sequence containing None (e.g.
"(None,)"). Since it turns out this behavior is useful when we
are reading in PHA-2 data, since we want to be able to treat a
missing column as a bunch of None values that we can just pass
through to the DataPHA instance, we add back in support where
we need this.

There is more clean up that could be done here, but I really
don't feel the need at this time.

Also addresses review comments by @hamogu