utils: replace func_counter by FuncCounter
Mark func_counter as deprecated and replace it with FuncCounter.

Rather than use a decorator to support counting the number of
times a function is called, use a specific class. This has the
advantage that

- the "nfev" attribute is kept with the function symbol, rather
  than being a separate variable

- we don't end up using a mutable list to store the value, which
  makes the logic cleaner

As part of this there is some apparently-unrelated code clean
up and the removal of a wrapper class that stored input/output
values for each iteration but that we never end up using (it
appears to have only been used for devlopmemt/testing based on
some commented-out print statements that are also removed in this
commit).

There is a surprising amount of change needed to accomodate this,
but it's all in what I claim are "specialised" bits of code that
a user is unlikely to be using.

There are also some follow-up questions that I mention in the
code, as comments, but they are not relevant here.