CI: bump the checkout actions version
https://github.blog/changelog/2023-09-22-github-actions-transitioning-from-node-16-to-node-20/

says we need to update, so do so.

There is the question of whether we change to

 - actions/checkout@v4
 - actions/checkout@v4.1.1
 - actions/checkout@v4.1.4

(v 4.1.4 is the latest). I have gone with the middle version since
v4.1.2 has been reported to be problematic - see
https://github.com/actions/checkout/releases/tag/v4.1.2 - so let's
be safe. I went with a specific version just because

a) the old code used a specefic version
b) the ability to pin the version couldbe useful

I note that we do not use much from the checkout action so we should
be okay just using a "stable" version.