opt: move call-back to top-level class  WIP
Some of the following needs reworking!

There are changes to the Opt class to move logic out into separate
classes, and to take time to rework the bounds checking (same bit of
code). Previously the code had allowed the bounds to be unspecified,
and - if that happened - they would be replaced by all -np.inf or all
np.inf. Now, the code is not complex, but

a) we never actually use it (i.e. we always send in limits)
b) the way the code was written it would probably fail if
   xmin was None - since the code also relied on xmin to
   determine the number of elements

So we remove this bit of logic.

There is some follow-on typing statements and use if super() to call
the super class, but hopefully nothing too excessive.

A number of places where we had trouble using multiprocessing with
non-fork methods (due to the inability for ForkingPickler to pickle
local symbols/routines) has been addressed by moving this code into
top-level callable classes.

Some of the code has been simplified, as there were functions that
called functions that called functions which could be converted into a
single callback class, as well as related code clean up (grid search).