estmethods: move call-backs into top-level classes.  WIP
This allows some code to be shared between the Confidence and
Projection classes (the FitCallback class) which wasn't obvious
from just looking at the code.

These changes are intended to allow the multiprocessing code to
work with non-fork methods - i.e. to allow the code to be pickled
by ForkingPickler.

There is also some changes to use super() rather than hard-code
the super-class name, change numpy to np, and replace the use
of a "default mutable argument" by changing statkwargs to default
to None rather than {}.