estmethods: move call-backs into top-level classes.  WIP
Need to check these comments

This allows some code to be shared between the Confidence and
Projection classes (the FitCallback class) which wasn't obvious
from just looking at the code.

These changes are intended to allow the multiprocessing code to
work with non-fork methods - i.e. to allow the code to be pickled
by ForkingPickler.

[hopefully the following has been moved to earlier commits]
There is also some changes to use super() rather than hard-code
the super-class name, change numpy to np.