fit: minor code cleanup
I had noted a while ago that StatInfoResults was a subclass of
NoNewAttributesAfterInit but did not initialize it during it's
initialization. I have therefore decided to change this to use
the __slots__ technique to mark this as "no new attributes" -
see issue #2008.

Rather than import several symbols from numpy just use the full
name (well, we import numpy as np so this means adding "np." before
the symbols) just so that readers know where these functions come
from.

Recent changes to the model class mean that we can replace

   [p for p in foo.get_thawed_pars()]

with

   foo.get_thawed_pars()

This is likely a hold-over from the original code where there was
no get_thawed_pars method so we had to select the parameters to
return.

There is some attempt to add typing rules for the code that is
being changed.