fit: rework output handling in the fit callback  WIP
This is a change to the API but it's a "definitely internal" change,
as it changes the _get_callback method. The main change is to move
the creation of the (optional) output file from the callback to
the code calling the callback. This means that the callback is
very-slightly cleaner as it doesn't have the responsibility for
creating a file handle.

It would be nice to wrap the file handle in a `with` block, to
support automatic clean up, but this is tricky given that it is
an optional behaviour.

There is also a little bit of code clean up in that we no longer
unpack a tuple into a list, change an element, and then pack it
back into a tuple. We now explicitly decompose the tuple, as we
know the size and fields we expect, and then reconstruct it. This
also makes it more obvious that we can use some of these fields
in the "write out the final fit results to the optional output
file" code.