utils: move is-subclass to central location
We have several variants of this routine so centralize it in
sherpa.utils. The different versions don't exactly match but
assume that the common version is sufficient (these tests are
not particularly "pythonic" but I do not want to try and remove
them as that would be a lot of churn).