IO: make sure the backend is always set
In the code it's okay to set backend to be None and then replace it,
but for type checkers that means that it thinks backend has a type
like Optional[module] rather than module, which means that whenever
backend is used the typing code wants you to know that backend could
be None. We can fix this easily, at the expense of losing
documentation of the backend variable.