IO: update typing rules in the pyfits and pycrates backends
This adds typing rules for more functions and moves to using
smoe aliases for common types. The modern way of doing this
is to use type, but that requires Python 3.12, or TypeAlias,
which requires Python 3.10, and we still support Python 3.9.

The attempt is to match dummy_backend.py for the routines but
there are some differences and it does not feel like we should
change them at the moment (e.g. the default values of the
fix_type and make_copy arguments).