IO: add typing rules to most of sherpa.io
There is some code change here, to try and appease mypy since we
do not explicitly state the type of the column data, and that
throws things off a bit. There is also the fact that checks like
numpy.iterable do not seem to match well with the typing approach
I have been using (preferring to use Sequence to represent an
iterable) since numpy.iterable tells mypy we have an Iterable[Any]
which apparently does not match Sized which len requires. So, I
have re-worked some code checks to avoid so many internal checks
and just rely on numpy creating a 2D (or larger) array, and catch
any conversion errors as well as a size check to try and match
the previous checks.