TEST: add corner case testing of sherpa.io and astro backend
Try to improve coverage and ensure existing behavior is tested.
In doing this, try to slightly improve the code. There is a balance
between "catching errors nicely" and "not breaking code", since
some users could (perhaps) be taking advantage of the current
behaviour, which is not always obvious. For instance

- write_arrays is expected to be sent a sequence of sequences,
  but you can send it just a sequence and each element is then
  considered a single column

- get_ascii_data only checks whether the input file lists too many
  column names for the data when colkeys is set

So we want to add enough tests to just cover the existing behaviour,
and to try and improve coverage.

However, in doing this I have found that numpy 1.24 changed the
asarray behavior: prevously it would create a "ragged" array
(i.e. allow each row to have a different number of columns) but
it would complain about it. From 1.24 it now errors out (unless
you force the output type to be object). There are several places
where we have code that may or may not assume we have a n by m
grid of data, and the code to do so (e.g. whether we need to be
explicit and check or just let numpy error out) therefore depends
on the numpy version in use. So the code attempts to be explicit
(i.e. check itself) first, so we can use pre and post numpy 1.24
(although numpy 1.23 is soon going to be marked end-of-life).