IO: try to standardize basic I/O error handling
Recent changes have led to different errors being raised in
some "basic" error condition cases (e.g. sending in a directory
rather than a file name). So we attempt to catch some lower-level
errors and make then an IOErr. This does not guarantee a consistent
error message, but hopefully it's more-likely to be an IOerr
rather than an OSError. Of course, this does lead to a subtle
change in behaviour when an error occurs - e.g. the error
raised may change and because IOErr does not derive from OSError
it could break some downstream software.

There is some attempt to use "standard" ways of accessing HDU
data on the pyfits side (e.g. accessing via .fileinfo() rather
than _file).