stats: add typing statements
We rely on passing around ndarray's rather than allowing the
input to be something like Sequence[SupportsFloat] | np.ndarray
just to make things easier to read. This may need to be relaxed
at some point.

The typing of the "calculate the stats or errors" routines is
not as good as we could like, in part because

- since many of the routines are defined in an extension module
  then Python doesn't ahve access to them / thinks they are
  "(...)", so it's hard to write something like a Protocol to
  represent them.

- the calling sequence depends on the statistic, as some of the
  more "complex" ones add different arguments in a less than
  principled manner

- the UserStat class also adds some complexity

The tests_stats.py file has seen a relatively-significant
internal rework to try and match the typing rules. It extends
the UserStat class but not in a way that matches the UserStat
interface: it should probably just extend Likelihood directly
but I have not gone that far as it wouldn't "improve" the
testing done.