stats: minor internal change to the UserStat class
Previously we had been using two variables to represent a
function: one for the function itself, and one for whether it
had been set. This has been replaced by making the function
optional (rather than it defaulting to 'lambda x: None'), which
lets us check the function itself for if it is defined, and
it simplifies the state handling (although it is not clear
if we really need to worry about that any more, I have left it
in just in case).