sherpa: add typing statements
These statements did find an error in one of the error handling
cases (this has already been fixed).

The JSON-like return value (using a dict to send around success
or error data) could be turned into an actual type or dataclass,
as it is very basic, but it does not seem worth it.