Add an explicit code on conduct specific to this project
This is a fairly standard software project doe of conduct that's been adopted by many projects in the community. I reviewed the SAO, CfA, and MIT specific codes, but they all don't really apply here, as they typically center on physical visitors to our facilities. Also, it is our long-term goal to make Sherpa a viable source project by itself and not just one of many scripts provided by the CXC. Thus, I think it's better to start with a CoC that's independent of our respective institutions.
However, as a start, I still put "contact the help desk" as the method to report CoC concerns to project leadership.