Add a security reporting policy
This text assumes that someone with "admin" permission in the repro (probably Warren?) activates github private security reporting, which wont' hurt: https://docs.github.com/en/code-security/security-advisories/working-with-repository-security-advisories/configuring-private-vulnerability-reporting-for-a-repository

If that form is activated, all repro admins will get an email and see the report in the "security" tab on the repro. They can then decide how to continue, e.g. share the report with other developers as needed.