xspec: allow Path for an XSPEC table model filename
Allow the filename argument to read_xstable_model to be a Path
object as well as a str. To support typing this code I changed
the creation of the XSTableModel so that the call is explicit
(in matching the data that has just been read in to the
constructor call) rather than just using *cols.

There should be no functional change in behavior here.