docs: create plots at build time for simulation
Make the code in the simulation tests

- testable
- create the plots rather than hard-code them

Note that not all the plot examples have been updated as they
require "extra knowledge" - e.g. what the arf or model expression
is meant to be.

This change means that we need the test data available when we
build Sherpa. The code will report when the data directory is
not found, but it will also create a number of follow-on errors
which is not ideal. Having the data directory available for the
RTD build is awkward because the build installs Sherpa as an
actual package (not as a development package) which means that
we can not just install the packae as a git submodule, since
those directories do not get installed by the Sherpa installation.
I have gone with the approach of installing the HEAD of the
sherpa/sherpa-test-data account, since we may need data newer
than the last release (and we don't have an easy way to install
the latest release in a generic manner). It **does** mean that
any PR that uses new or updated data files and needs them in
the documentation will fail the RTD GitHub builds until the
sherpa-test-data account is updated.