optfcts: internal code clean up
Add some comments on code that can probably be changed, as well
as avoiding a potential issue where a variable can be an integer,
sequence, or ndarray and mypy gets confused. It makes it a bit
clearer that the finalsimplex argument - which I had documented
as lonly taking an integer but could actually take a sequence -
is not actually tested in the latter case.