optfcts: add some basic typing rules
This is by no means complete, and is an approximation (e.g.
the types of array inputs is not really

    Sequence[SupportsFloat] | np.ndarray

but it is hopefully close enough to be useful.

There is likely some tension between using float vs SupportsFloat,
but that can be addressed when necessary as the difference isn't
huge for our use cases.