optfcts: internal clean up of neldermead
The neldermead routine has been documented as accepting None or
an array for the step argument, but the code actually

- allowed for a scalar
- allowed for step to be a sequence but of the wrong size

The former is not tested, and the latter is just bad design for
such a low-level routine (i.e. it should error out in this case),
so clean up the routine to remove this extra complexity. It
now follows the same approach as minim().