optfcts: internal code clean up for optfcts
Use standard NumPy element-wise addition/multiplication rather
than doing it manually with list/map and a lambda expression.

The parallel_map routine is also guaranteed to return a list,
so there's no need to cast the return value (also, we don't
actually care that it's a list, just that it can be treated
as a Sequence in this context).