optmethods: does this clean up clang warnings
In issue #1542 we noted that clang was creating a lot of
warning messages from the use of ParVal and its use of virtual
functions. There are a number of ways to address this but this,
if it works, is very simple. We are not planning to extend the
C++ code so marking the class as final should not be a problem.