fit: i/o changes
We know explicitly allow the output file for fit calls to be
a Path object (previously we allowed it but not intentionally)
and also allow a file-handle like object to be used. This then
lets you say

    out = String()
    fit(outfile=out)
    txt = out.getvalue()

without having to create an on-disk file.

In doing this the internals of the IterFit class were changed,
and the callback routine it uses was converted from a local
definition to a top-level class. This change I think better
separates the code logic, but is also to help address
issue #2007 (where the multiprocessing code doesn't seem to like
our use of local function definitions when using anything but
the "fork" method).

As part of this work some basic typing statements have been
added. There is some fun in how best to indicate the type of
a "file handle": the easy solution is typing.TextIO but this
is "too large" as all we need are write() and close() methods,
so we use a Protocol. See the discussion at
https://github.com/python/typing/discussions/829
Now, this may not be ideal - as we don't currently export the
protocol so how will it work with the documentation - but it
would be easy to change to use TextIO.