fit: internal code change
Clarify the internal fitting code, where a tuple was being
converted to a list to be able to edit an element. We now explicitly
decompose the tuple, so it is more obvious what is going on.

I had hoped to be able to simplify this routine even more, but
the difference between the statistic object stored locally and
the one stored in the callback means it is not as simple as I
had hoped.