data: add minimal typing
This is to support typing sherpa.fit.

We have some things to think about here

- the general "how best to specify a numeric array" issue we have
  throughout the code base (there are some places where we assume
  model evaluation returns a ndarray but there's no requirement
  we do so, and so should we just make sure we call np.asarray
  in these cases?)

- do we want to say you can only send Model classes, or should
  we support a more Protocol-based approach so we say "something
  that behaves like a Model"? The latter is more pythonic, but
  only in places where we only need the "callable" part of a model

  As we are not concentrating on the model classes here we just
  want something to say "hey, this looks like a model", and we can
  refine this when the model code is looked at.

Unfortunately neither mypy or pyright accepts sherpa/data.py after
these changes, but this is in part thanks to the difficulty of typing
our code, and other parts it's my lack of understanding, and other
parts the lack of maturity of the Python typing ecosystem.

The typing code suggests that we could tighten-up things a bit - e.g.
more sever handling of "axis not set" or "axis has been masked out" -
but this is for a later PR.  There are also places where I have
decided not to add types as it would just get confusing at this point.

It is also obvious that many of our sub-classes completely
redefine the API of a method, which can make it hard to follow.
It may be worth looking at the existing class hierarchy, as
to_plot should probably

a) return a structured data type, rather than a tuple
b) be moved to Data

This adds a sherpa.utils.types module which is used to have a
place where common useful types are placed. This module is
intended to be an "internal" module - that is, not used by
external code - so stability requirements are limited (i.e.
things can change here often).